refactor: update Bolt Core to point to util functions now living in @bolt/element
